move expression tests to their own class
