+++ corehq/apps/userreports/tests/test_expressions.py
@@ -1,0 +1,57 @@
+from django.test import SimpleTestCase
+from corehq.apps.userreports.exceptions import BadSpecError
+from corehq.apps.userreports.expressions.factory import ExpressionFactory
+from corehq.apps.userreports.expressions.getters import NestedDictGetter, DictGetter
+
+
+class ExpressionFromSpecTest(SimpleTestCase):
+
+    def test_invalid_type(self):
+        with self.assertRaises(BadSpecError):
+            ExpressionFactory.from_spec({
+                'type': 'not_a_valid_type',
+            })
+
+    def test_property_name_expression(self):
+        getter = ExpressionFactory.from_spec({
+            'type': 'property_name',
+            'property_name': 'foo',
+        })
+        self.assertEqual(DictGetter, type(getter))
+        self.assertEqual('foo', getter.property_name)
+
+    def test_property_name_no_name(self):
+        with self.assertRaises(BadSpecError):
+            ExpressionFactory.from_spec({
+                'type': 'property_name',
+            })
+
+    def test_property_name_empty_name(self):
+        with self.assertRaises(BadSpecError):
+            ExpressionFactory.from_spec({
+                'type': 'property_name',
+                'property_name': None,
+            })
+
+    def test_property_path_expression(self):
+        getter = ExpressionFactory.from_spec({
+            'type': 'property_path',
+            'property_path': ['path', 'to', 'foo'],
+        })
+        self.assertEqual(NestedDictGetter, type(getter))
+        self.assertEqual(['path', 'to', 'foo'], getter.property_path)
+
+    def test_property_path_no_path(self):
+        with self.assertRaises(BadSpecError):
+            ExpressionFactory.from_spec({
+                'type': 'property_path',
+            })
+
+    def test_property_path_empty_path(self):
+        for empty_path in ([], None):
+            with self.assertRaises(BadSpecError):
+                ExpressionFactory.from_spec({
+                    'type': 'property_path',
+                    'property_path': empty_path,
+                })
+
