+++ corehq/apps/userreports/tests/test_getters.py
@@ -75,55 +75,3 @@ class TransformedGetterTest(SimpleTestCase):
         getter = TransformedGetter(self.base_getter, lambda x: '{}-transformed'.format(x))
         self.assertEqual('bar-transformed', getter({'foo': 'bar'}))
         self.assertEqual('1-transformed', getter({'foo': 1}))
-
-
-class ExpressionFromSpecTest(SimpleTestCase):
-
-    def test_invalid_type(self):
-        with self.assertRaises(BadSpecError):
-            ExpressionFactory.from_spec({
-                'type': 'not_a_valid_type',
-            })
-
-    def test_property_name_expression(self):
-        getter = ExpressionFactory.from_spec({
-            'type': 'property_name',
-            'property_name': 'foo',
-        })
-        self.assertEqual(DictGetter, type(getter))
-        self.assertEqual('foo', getter.property_name)
-
-    def test_property_name_no_name(self):
-        with self.assertRaises(BadSpecError):
-            ExpressionFactory.from_spec({
-                'type': 'property_name',
-            })
-
-    def test_property_name_empty_name(self):
-        with self.assertRaises(BadSpecError):
-            ExpressionFactory.from_spec({
-                'type': 'property_name',
-                'property_name': None,
-            })
-
-    def test_property_path_expression(self):
-        getter = ExpressionFactory.from_spec({
-            'type': 'property_path',
-            'property_path': ['path', 'to', 'foo'],
-        })
-        self.assertEqual(NestedDictGetter, type(getter))
-        self.assertEqual(['path', 'to', 'foo'], getter.property_path)
-
-    def test_property_path_no_path(self):
-        with self.assertRaises(BadSpecError):
-            ExpressionFactory.from_spec({
-                'type': 'property_path',
-            })
-
-    def test_property_path_empty_path(self):
-        for empty_path in ([], None):
-            with self.assertRaises(BadSpecError):
-                ExpressionFactory.from_spec({
-                    'type': 'property_path',
-                    'property_path': empty_path,
-                })
